realtime ping now worked
